+++ components/layout/lib.rs
@@ -46,6 +46,7 @@ pub mod layout_debug;
 pub mod block;
 pub mod construct;
 pub mod context;
+pub mod display_list_builder;
 pub mod floats;
 pub mod flow;
 pub mod flow_list;
